update major validate and everthing else, stable

+++ app/Http/Controllers/Master/UsersController.php
@@ -22,7 +22,7 @@ class UsersController extends Controller
     public function index()
     {
         $users = User::all(); // Fetch all users
-        return view('master.users.index', compact('users')); // return a view to display the users
+        return view('users.index', compact('users')); // return a view to display the users
     }
 
     /**
@@ -33,7 +33,7 @@ class UsersController extends Controller
         
         $positions = Position::restricted()->get(); // Fetch all positions
         $department = MasterManagementService::getDepartmentListOptions()->get();
-        return view('master.users.create', compact('positions','department')); // return the form view with the positions
+        return view('users.create', compact('positions','department')); // return the form view with the positions
     }
 
     /**
@@ -77,7 +77,7 @@ class UsersController extends Controller
         if (!$user) {
             abort(404);
         }
-        return view('master.users.show', compact('user')); // Return view to display user details
+        return view('users.show', compact('user')); // Return view to display user details
     }
 
     /**
@@ -91,7 +91,7 @@ class UsersController extends Controller
         }
         $positions = Position::restricted()->get(); // Gunakan scope restricted
         $department = MasterManagementService::getDepartmentListOptions()->get();
-        return view('master.users.edit', compact('user', 'positions','department')); // Return form view with user and positions
+        return view('users.edit', compact('user', 'positions','department')); // Return form view with user and positions
     }
 
     /**
@@ -127,7 +127,7 @@ class UsersController extends Controller
     //         return redirect()->back()->withInput()->with('error', 'Gagal update user: ' . $e->getMessage());
     //    }
         MasterManagementService::storeLogActivity('update-user',$user->id,$user->name);
-        return redirect()->route('master.users.index')->with('success', 'User updated successfully');
+        return redirect()->route('users.index')->with('success', 'User updated successfully');
     } catch (\Exception $e) {
         return redirect()->back()->withInput()->with('error', 'Gagal update user: ' . $e->getMessage());
     }
@@ -139,13 +139,18 @@ class UsersController extends Controller
     public function destroy(string $id)
     {
         $user =User::where('id',$id)->restricted()->first();
+        $current_user = AuthService::currentAccess();
+        // dd($user->position->roles[0]->name);
+        if ($current_user['role'] == 'admin' && $user->position->roles[0]->name == 'admin') {
+            return redirect()->back()->withInput()->with('error', 'Forbidden! Cannot Delete This User');
+        }
         if (!$user) {
             abort(404);
         }
         $user->delete(); // Delete user
         MasterManagementService::storeLogActivity('delete-user',$user->id,$user->name);
 
-        return redirect()->route('master.users.index')->with('success', 'User deleted successfully');
+        return redirect()->route('users.index')->with('success', 'User deleted successfully');
     }
 
     public function editProfile()
@@ -153,7 +158,7 @@ class UsersController extends Controller
         $user = User::find(AuthService::currentAccess()['id']);
         $positions = Position::all();
         $department = Department::all();
-        return view('master.users.edit-profile', compact('user', 'positions', 'department'));
+        return view('users.edit-profile', compact('user', 'positions', 'department'));
     }
 
     /**
